pyproject.toml
[tool.poetry]
name = "tivoli"
version = "0.1.0"
description = "JWT token validator"
authors = ["Eugen Ciur <eugen@papermerge.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
python-jose = "^3.3.0"
uvicorn = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
